+++ src/app/cart/components/cart-list/cart-list.component.ts
@@ -1,6 +1,7 @@
 import { Component, ViewChild, ElementRef, Renderer2 } from '@angular/core';
 import { OrderService } from 'src/app/order/services/order.service';
 import { ProductModel } from 'src/app/product/models/product.model';
+import { CartDTOModel } from '../../models/cart-dto.model';
 
 @Component({
   selector: 'app-cart-list',
@@ -12,17 +13,18 @@ export class CartListComponent {
   @ViewChild('orderedP')
   orderedP: ElementRef;
 
-  orderedProducts: Array<ProductModel>;
-
   constructor(
     private orderService: OrderService,
     private renderer: Renderer2
   ) {
-    this.orderedProducts = orderService.getSelectedProducts();
   }
 
   onDelete(product: ProductModel) {
     this.orderService.removeProduct(product);
+  }
+
+  onChangeQuantity(cartDTO: CartDTOModel) {
+      this.orderService.updateQuantity(cartDTO);
   }
 
   onWheel(event: any) {
@@ -31,6 +33,10 @@ export class CartListComponent {
     } else {
       this.renderer.setStyle(this.orderedP.nativeElement, "font-size", "15px");
     }
+  }
+
+  getOrderedProducts(): ProductModel[] {
+    return this.orderService.getSelectedProducts();
   }
 
   getOrdersQuanity(): number {