+++ src/app/cart/models/cart-dto.model.ts
@@ -1,0 +1,8 @@
+import { ProductModel } from 'src/app/product/models/product.model';
+
+export class CartDTOModel {
+  constructor(
+    public product: ProductModel,
+    public quantity: number
+  ) { }
+}